+++ src/Mosaic.php
@@ -28,11 +28,18 @@ class Mosaic
     private $target;
 
     /**
-     * Log container
+     * HTML table container
      *
      * @var array<string>
      */
-    private $log;
+    private $htmlTable;
+
+    /**
+     * Images information container
+     *
+     * @var array
+     */
+    private $images;
 
     //-----------------------------------------------------------------------------------------------------------------
     // PUBLIC METHODS
@@ -46,21 +53,20 @@ class Mosaic
      *
      * @return void
      */
-    public function __construct(string $filename, bool $enableLog = false)
+    public function __construct(string $filename)
     {
-        $this->log = [
-            'isEnabled' => $enableLog,
-            'contents'  => [],
-        ];
+        $this->htmlTable = [];
+        $this->images    = [];
 
         $this->source = new Histogram();
         $this->source->loadFromFile($filename);
 
-        $this->target = imagecreatetruecolor(
-            $this->source->getDetails()['size']->width,
-            $this->source->getDetails()['size']->height
+        $this->target = \imagecreatetruecolor(
+            $this->source->getDetails()['size']['width'],
+            $this->source->getDetails()['size']['height']
         );
-        imagealphablending($this->target, false);
+        \imagealphablending($this->target, false);
+        \imagesavealpha($this->target, true);
     }
 
     //-----------------------------------------------------------------------------------------------------------------
@@ -68,66 +74,261 @@ class Mosaic
     /**
      * @access public
      *
-     * @param  string $filename
-     * @param  int $cols
-     * @param  int $rows
+     * @param  string $pattern
+     * @param  string $pathCache
+     * @param  bool $reloadCache
      *
      * @return void
      */
-    public function create(string $filename, int $cols, int $rows) : void
+    public function loadImages(string $pattern, string $pathCache, bool $reloadCache = false) : void
     {
-        $partialWidth  = floor($this->source->getDetails()['size']->width / $cols);
-        $partialHeight = floor($this->source->getDetails()['size']->height / $rows);
+        $this->images = Histogram::processImages($pattern, $pathCache, $reloadCache);
+    }
 
-        if ($this->log['isEnabled']) {
-            $this->log['contents'][] = 'P1(X, Y) - P2(X, Y) - COLOR';
-            $this->log['contents'][] = '----------------------------------------';
-        }
+    //-----------------------------------------------------------------------------------------------------------------
+
+    /**
+     * @access public
+     *
+     * @param  int $cols
+     * @param  int $rows
+     * @param  bool $withImages
+     *
+     * @return array
+     */
+    public function create(int $cols, int $rows, bool $withImages = false) : array
+    {
+        $sectionWidth  = (int) floor($this->source->getDetails()['size']['width'] / $cols);
+        $sectionHeight = (int) floor($this->source->getDetails()['size']['height'] / $rows);
 
         foreach (range(1, $rows) as $row) {
             foreach (range(1, $cols) as $col) {
-                $p1x = (int) (($col - 1) * $partialWidth);
-                $p1y = (int) (($row - 1) * $partialHeight);
-                $p2x = (int) ($col * $partialWidth);
-                $p2y = (int) ($row * $partialHeight);
-
-                $this->source->setArea($p1x, $p1y, $p2x, $p2y);
+                $this->source->setArea(
+                    (int) (($col - 1) * $sectionWidth),
+                    (int) (($row - 1) * $sectionHeight),
+                    (int) ($col * $sectionWidth),
+                    (int) ($row * $sectionHeight)
+                );
 
                 $hexColor = $this->source->getAverageColor();
 
-                if ($this->log['isEnabled']) {
-                    $this->log['contents'][] = sprintf('%04d, %04d - %04d, %04d - %s',
-                        $p1x,
-                        $p1y,
-                        $p2x,
-                        $p2y,
-                        $hexColor
-                    );
+                $this->htmlTable[$row][$col] = [
+                    'p1x'   => $this->source->getArea()['p1']->x,
+                    'p1y'   => $this->source->getArea()['p1']->y,
+                    'p2x'   => $this->source->getArea()['p2']->x,
+                    'p2y'   => $this->source->getArea()['p2']->y,
+                    'color' => $hexColor,
+                ];
+
+                if (count($this->images) && $withImages) {
+                    $this->fillAreaWithImage($hexColor, $sectionWidth, $sectionHeight);
+
+                    continue;
                 }
 
-                $rgbaColor = ColorHex::toRgba($hexColor);
-
-                imagefilledrectangle(
-                    $this->target,
-                    $p1x,
-                    $p1y,
-                    $p2x,
-                    $p2y,
-                    imagecolorallocate(
-                        $this->target,
-                        (int) $rgbaColor['r'],
-                        (int) $rgbaColor['g'],
-                        (int) $rgbaColor['b']
-                    )
-                );
+                $this->fillAreaWithColor($hexColor, $sectionWidth, $sectionHeight);
             }
         }
 
-        imagepng($this->target, $filename);
+        return [
+            $this->saveAsPng($rows, $cols, $sectionWidth, $sectionHeight),
+            $this->saveAsHtml($rows, $cols, $sectionWidth, $sectionHeight),
+        ];
+    }
 
-        if ($this->log['isEnabled']) {
-            file_put_contents('mosaic.log', implode(PHP_EOL, $this->log['contents']));
+    //-----------------------------------------------------------------------------------------------------------------
+    // PRIVATED METHODS
+    //-----------------------------------------------------------------------------------------------------------------
+
+    /**
+     * @access private
+     *
+     * @param  string $hexColor
+     *
+     * @return array
+     */
+    private function findClosestImage(string $hexColor) : array
+    {
+        $currentDistance = PHP_INT_MAX;
+        $image           = null;
+
+        foreach ($this->images as $entry) {
+            $distance = ColorHex::distanceCie76($hexColor, $entry['colors']['average']);
+
+            if ($distance < $currentDistance) {
+                $currentDistance = $distance;
+                $image           = $entry;
+            }
         }
+
+        return $image;
+    }
+
+    //-----------------------------------------------------------------------------------------------------------------
+
+    /**
+     * @access private
+     *
+     * @param  string $hexColor
+     * @param  int $sectionWidth
+     * @param  int $sectionHeight
+     *
+     * @return array
+     */
+    private function fillAreaWithImage(string $hexColor, int $sectionWidth, int $sectionHeight) : void
+    {
+        $closest = $this->findClosestImage($hexColor);
+
+        $source = \imagecreatefromstring(
+            file_get_contents(
+                $closest['details']['dirname'] .'/'. $closest['details']['basename']
+            )
+        );
+
+        \imagecopyresampled(
+            $this->target,
+            $source,
+            $this->source->getArea()['p1']->x,
+            $this->source->getArea()['p1']->y,
+            0,
+            0,
+            $sectionWidth,
+            $sectionHeight,
+            $closest['details']['size']['width'],
+            $closest['details']['size']['height']
+        );
+
+        \imagedestroy($source);
+    }
+
+    //-----------------------------------------------------------------------------------------------------------------
+
+    /**
+     * @access private
+     *
+     * @param  string $hexColor
+     * @param  int $sectionWidth
+     * @param  int $sectionHeight
+     *
+     * @return array
+     */
+    private function fillAreaWithColor(string $hexColor, int $sectionWidth, int $sectionHeight) : void
+    {
+        $rgbaColor = ColorHex::toRgba($hexColor);
+
+        \imagefilledrectangle(
+            $this->target,
+            $this->source->getArea()['p1']->x,
+            $this->source->getArea()['p1']->y,
+            $this->source->getArea()['p2']->x,
+            $this->source->getArea()['p2']->y,
+            \imagecolorallocate(
+                $this->target,
+                (int) $rgbaColor['r'],
+                (int) $rgbaColor['g'],
+                (int) $rgbaColor['b']
+            )
+        );
+    }
+
+    //-----------------------------------------------------------------------------------------------------------------
+
+    /**
+     * Save as PNG
+     *
+     * @access private
+     *
+     * @param  int $rows
+     * @param  int $cols
+     * @param  int $sectionWidth
+     * @param  int $sectionHeight
+     *
+     * @return string
+     */
+    private function saveAsPng(int $rows, int $cols, int $sectionWidth, int $sectionHeight) : string
+    {
+        $filename = sprintf(
+            '%s/%s-%dx%d.png',
+            $this->source->getDetails()['dirname'],
+            $this->source->getDetails()['filename'],
+            $rows,
+            $cols
+        );
+
+        \imagepng($this->target, $filename);
+
+        return $filename;
+    }
+
+    //-----------------------------------------------------------------------------------------------------------------
+
+    /**
+     * Save as HTML
+     *
+     * @access private
+     *
+     * @param  int $rows
+     * @param  int $cols
+     * @param  int $sectionWidth
+     * @param  int $sectionHeight
+     *
+     * @return string
+     */
+    private function saveAsHtml(int $rows, int $cols, int $sectionWidth, int $sectionHeight) : string
+    {
+        $filename = sprintf(
+            '%s/%s-%dx%d.html',
+            $this->source->getDetails()['dirname'],
+            $this->source->getDetails()['filename'],
+            $rows,
+            $cols
+        );
+
+        $getTableBody = function () : string {
+            $tbody = [];
+            foreach ($this->htmlTable as $row => $aux) {
+                $tr = [];
+
+                $tr[] = '<tr>';
+                foreach ($aux as $col => $data) {
+                    $tr[] = sprintf(
+                        '<td data-section="(%d,%d) - (%d,%d)" style="background-color: %s"></td>',
+                        $data['p1x'],
+                        $data['p1y'],
+                        $data['p2x'],
+                        $data['p2y'],
+                        $data['color']
+                    );
+                }
+                $tr[] = '</tr>';
+
+                $tbody[] = implode(PHP_EOL, $tr);
+            }
+
+            return implode(PHP_EOL, $tbody);
+        };
+
+        $template = file_get_contents(__DIR__ .'/Mosaic.tpl.php');
+
+        file_put_contents($filename, strtr($template, [
+            '__TITLE__'   => sprintf(
+                'Source [ %s ] - Mosaic [ %d x %d ]',
+                $this->source->getDetails()['basename'],
+                $rows,
+                $cols
+            ),
+            '__CAPTION__' => sprintf(
+                'Source [ %s ] - Mosaic [ %d x %d ]',
+                $this->source->getDetails()['basename'],
+                $rows,
+                $cols
+            ),
+            '__TBODY__'       => $getTableBody(),
+            '__CELL_WIDTH__'  => $sectionWidth,
+            '__CELL_HEIGHT__' => $sectionHeight,
+        ]));
+
+        return $filename;
     }
 
     //-----------------------------------------------------------------------------------------------------------------