+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All Notable changes to `mosaic` will be documented in this file
 
+## 1.0.1 - 2020-04-24
+
+- Added extra HTML output in where `mosaic` is a `<table>` DOM element.
+- Allowed to generate the `mosaic` with images instead of plain colors.
+
 ## 1.0.0 - 2020-04-22
 
 - First release